Break out dash/open_meteo.py
