+++ dash/weather_api.py
@@ -8,7 +8,7 @@ WEATHER_API_KEY = os.environ["WEATHER_API_KEY"]
 assert len(WEATHER_API_KEY) == 31
 
 
-def weather_api_fetch_forecast(zip_code: str, country_code: str = "US") -> dict:  # type: ignore
+def weather_api_fetch_forecast(zip_code: str, country_code: str = "US") -> dict[str, str]:
     params = {
         "key": WEATHER_API_KEY,
         "q": f"{zip_code},{country_code}",
@@ -17,7 +17,5 @@ def weather_api_fetch_forecast(zip_code: str, country_code: str = "US") -> dict:
         "hour": "1",  # Get hourly forecasts
     }
     response = httpx.get(WEATHER_API_BASE_URL, params=params)
-    print(response.status_code, type(response), 1)
-    print(response.text, 2)
     response.raise_for_status()
     return dict(response.json())
